the-architect/backend/backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
daphne==4.0.0
google-generativeai==0.3.0
PyPDF2==3.0.1
python-docx==0.8.11
requests==2.31.0
python-decouple==3.8
